requirements.txt
# REST API libraries
flask
python-dotenv
# Purview Python libraries
azure-identity 
azure-purview-catalog
azure-purview-scanning
purviewcli==0.1.34
# Python helpers
jmespath